+++ api/auth/v1/auth.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, Depends, Response
-from core.exceptions import ExceptionResponseSchema, DecodeTokenException
+from core.exceptions.responses import ExceptionResponseSchema
+from core.exceptions.token import DecodeTokenException
 from core.fastapi.dependencies.permission import AllowAll, PermissionDependency
 from core.fastapi_versioning import version
