+++ api/ingredient/v1/ingredient.py
@@ -2,7 +2,7 @@
 
 from typing import List
 from fastapi import APIRouter, Depends
-from core.exceptions import ExceptionResponseSchema
+from core.exceptions.responses import ExceptionResponseSchema
 from core.fastapi_versioning import version
 from core.fastapi.dependencies.permission import (
     AllowAll,
