+++ api/recipe/v1/recipe.py
@@ -1,7 +1,7 @@
 """Endpoints for recipe.
 """
 from fastapi import APIRouter, Depends
-from core.exceptions import ExceptionResponseSchema
+from core.exceptions.responses import ExceptionResponseSchema
 from core.fastapi.dependencies.user import get_current_user
 from core.fastapi_versioning import version
 from core.fastapi.dependencies.permission import (
