+++ app/recipe/services/recipe.py
@@ -3,8 +3,9 @@
 from typing import List, Dict
 from core.db.models import RecipeIngredient, Recipe, RecipeTag, User
 from core.db import Transactional
-from core.exceptions import RecipeNotFoundException, UserNotFoundException
 from core.exceptions.base import UnauthorizedException
+from app.recipe.exceptions.recipe import RecipeNotFoundException
+from app.user.exceptions.user import UserNotFoundException
 from app.ingredient.repository.ingredient import IngredientRepository
 from app.tag.repository.tag import TagRepository
 from app.tag.schemas import CreateTagSchema
