+++ app/tag/repository/tag.py
@@ -2,7 +2,7 @@
 The module contains a repository class that defines database operations for tags. 
 """
 
-from typing import List
+from typing import List, Type
 from sqlalchemy import select
 from core.db import session
 from core.db.models import Tag
@@ -11,6 +11,9 @@ from core.repository.base import BaseRepo
 
 class TagRepository(BaseRepo):
     """Repository for tag related database operations"""
+
+    def __init__(self):
+        super().__init__(Tag)
 
     async def create_tag(self, name: str, tag_type: str) -> Tag:
         """
