+++ app/user/services/user.py
@@ -10,11 +10,8 @@ from app.user.utils import generate_name, get_password_hash
 from app.user.repository.user import UserRepository
 from app.user.schemas.user import UpdateUserSchema
 from app.image.exceptions.image import FileNotFoundException
+from app.user.exceptions.user import UserNotFoundException, DuplicateUsernameException
 from core.db.models import User
-from core.exceptions import (
-    DuplicateUsernameException,
-    UserNotFoundException,
-)
 from core.db.session import session
